Reorder JS structure. Postcode AJAX rework. Install jquery via npm.
