+++ package.json
@@ -17,7 +17,7 @@
         "browser-sync-webpack-plugin": "^2.0.1",
         "cross-env": "^5.1",
         "gulp": "^4.0.2",
-        "jquery": "^3.2",
+        "jquery": "^3.4.1",
         "laravel-mix": "^4.0.16",
         "lodash": "^4.17.15",
         "popper.js": "^1.12",
