+++ resources/js/scripts/year.js
@@ -1,0 +1,6 @@
+// Output time (year) dynamically
+$( document ).ready(function() {
+    var date = new Date();
+    var thisyear = date.getFullYear();
+    $('#thisYear').text(thisyear);
+});
